+++ chatinterface_server/internal/config.py
@@ -25,7 +25,7 @@ class ConfigMaker:
         self.log_dir: str = log_dir
         self.conf_dir: str = conf_dir
 
-    def create_logging_config(self, log_level: str, use_console: bool = False) -> dict:
+    def create_logging_config(self, log_level: str, use_console: bool = True) -> dict:
         handlers: dict = {
             'console': {
                 'class': 'logging.StreamHandler',
@@ -35,12 +35,15 @@ class ConfigMaker:
         }
         loggers: dict = {}
 
-        routes: set[str] = {'ws', 'auth', 'chats', 'database'}
-        for route_name in routes:
-            logfile: str = os.path.join(self.log_dir, f'{route_name}.log')
-            handler_name: str = f'chatinterface.handler.{route_name}'
+        log_targets: set[str] = {
+            'ws', 'auth', 'chats', 
+            'database', 'main'
+        }
+        for target_name in log_targets:
+            logfile: str = os.path.join(self.log_dir, f'{target_name}.log')
+            handler_name: str = f'chatinterface.handler.{target_name}'
 
-            logger_name: str = f'chatinterface.logger.{route_name}'
+            logger_name: str = f'chatinterface.logger.{target_name}'
             handler_list: list[str] = [handler_name]
 
             if use_console:
@@ -76,58 +79,16 @@ class ConfigMaker:
             'loggers': loggers
         }
 
-    def create_database_config(self) -> dict:
-        # todo: make env override database.json or only used for first setup
-        db_host: str = os.getenv('CHATINTERFACE_DB_HOST', 'localhost')
-        db_port: str = os.getenv('CHATINTERFACE_DB_PORT', '3306')
-
-        db_name: str = os.getenv("CHATINTERFACE_DB_NAME", 'chatinterface')
-        db_user: str = os.getenv("CHATINTERFACE_DB_USER", 'chatinterface')
-
-        db_password: str = os.getenv("CHATINTERFACE_DB_PASSWORD", '')
-        if not db_port.isdigit():
-            raise ValueError("database port is not a valid port number")
-        else:
-            db_port: int = int(db_port)
-
-        db_conf: dict = {
-            'mysql_config': {
-                'host': db_host,
-                'port': db_port,
-                'db': db_name,
-                'user': db_user,
-                'password': db_password,
-                'pool_size': 10,
-                'pool_name': 'chatinterface-mysql-pool',
-                'pool_reset_session': True
-            },
-            'executor_config': {
-                'max_threads': 10,
-                'thread_name_prefix': 'mysql-databaseThread-'
-            }
-        }
-        return db_conf
-
 
 class ConfigManager:
     def __init__(self, base_dir: str = DEFAULT_APP_DIR) -> None:
         if not base_dir:
             raise ValueError("application base directory required but is empty")
         if not os.path.isdir(base_dir):
-            raise ValueError("application base directory does not exist, have you created it?")
+            os.makedirs(base_dir, mode=0o755)
 
-        self.conf_dir: str = ''
-        self.log_dir: str = ''
-
-        self.setup_directories(base_dir)
-        self.conf_maker: ConfigMaker = ConfigMaker(self.log_dir, self.conf_dir)
-
-        main_config_file: str = os.path.join(self.conf_dir, 'main.json')
-        main_config: dict = load_or_create_config(main_config_file, DEFAULT_MAIN_CONFIG)  # noqa
-
-    def setup_directories(self, base_dir: str) -> None:
-        abs_base_dir: str = os.path.abspath(base_dir)
         # if two instances are being ran, version separates instances if base dir is same
+        abs_base_dir: str = os.path.abspath(base_dir)
         ver_base_dir: str = os.path.join(abs_base_dir, __version__)
 
         log_dir: str = os.path.join(ver_base_dir, 'logs')
@@ -137,9 +98,14 @@ class ConfigManager:
         os.makedirs(log_dir, mode=0o700, exist_ok=True)
 
         os.makedirs(conf_dir, mode=0o700, exist_ok=True)
-        self.log_dir: str = log_dir
 
+        self.log_dir: str = log_dir
         self.conf_dir: str = conf_dir
+
+        self.conf_maker: ConfigMaker = ConfigMaker(self.log_dir, self.conf_dir)
+        main_config_file: str = os.path.join(self.conf_dir, 'main.json')
+
+        main_config: dict = load_or_create_config(main_config_file, DEFAULT_MAIN_CONFIG)  # noqa
 
     def setup_logging(self, log_level: str = "INFO") -> None:
         log_config_file: str = os.path.join(self.conf_dir, 'log.json')
@@ -147,10 +113,3 @@ class ConfigManager:
 
         log_config: dict = load_or_create_config(log_config_file, default_logging_config)
         logging.config.dictConfig(log_config)
-
-    def get_database_config(self) -> dict:
-        db_config_file: str = os.path.join(self.conf_dir, 'database.json')
-        default_db_config: dict = self.conf_maker.create_database_config()
-
-        db_config: dict = load_or_create_config(db_config_file, default_db_config)
-        return db_config