+++ chatinterface_server/models/__init__.py
@@ -1,0 +1,2 @@
+from . import ws, chats, common
+__all__: list[str] = ['ws', 'chats', 'common']