+++ chatinterface_server/models/chats.py
@@ -1,0 +1,6 @@
+from typing import Annotated
+from pydantic import BaseModel, Field
+
+class ComposeMessage(BaseModel):
+    recipient: Annotated[str, Field(max_length=20, description="Recipient username")]
+    message_data: str